app/requirements.txt
pandas 
openpyxl
mysql-connector-python